iCheck & begin update authen..

+++ msis-web/app/scripts/services/rest-service.js
@@ -1,24 +1,52 @@
 angular.module('webappApp')
-    .factory('RestService', ['$http', '$q', '$log', function($http, $q, $log) {
+    .factory('RestService', ['$http', '$q', '$log', '$rootScope', function($http, $q, $log, $rootScope) {
         function _get(uri) {
             $log.info('-> Get data from: ' + uri.substring(uri.lastIndexOf('/') + 1));
-            var deferred = $q.defer();
-            return $http.get(uri).then(successCallback, errorCallback);
+            // TODO: check signin/reg
+            if ($rootScope.currentUser == null) {
+                toastr.warning("Session was expired. Please signin again!", 'Msis-Web');
+                AuthenService.clearCredentials();
+                $location.path('/signin');
+            } else {
+                return $http.get(uri).then(successCallback, errorCallback);
+            }
+            return false;
         }
 
         function _post(uri, data) {
             $log.info('-> Post data to: ' + uri.substring(uri.lastIndexOf('/') + 1));
-            return $http.post(uri, data).then(successCallback, errorCallback);
+            if ($rootScope.currentUser == null) {
+                toastr.warning("Session was expired. Please signin again!", 'Msis-Web');
+                AuthenService.clearCredentials();
+                $location.path('/signin');
+            } else {
+                return $http.post(uri, data).then(successCallback, errorCallback);
+            }
+            return false;
         }
 
         function _put(uri, data) {
             $log.info('-> Put data to: ' + uri.substring(uri.lastIndexOf('/') + 1));
-            return $http.put(uri, data).then(successCallback, errorCallback);
+            if ($rootScope.currentUser == null) {
+                toastr.warning("Session was expired. Please signin again!", 'Msis-Web');
+                AuthenService.clearCredentials();
+                $location.path('/signin');
+            } else {
+                return $http.put(uri, data).then(successCallback, errorCallback);
+            }
+            return false;
         }
 
         function _delete(uri, id) {
             $log.info('-> Delete data from: ' + uri.substring(uri.lastIndexOf('/') + 1));
-            return $http.delete(uri + '/' + id).then(successCallback, errorCallback);
+            if ($rootScope.currentUser == null) {
+                toastr.warning("Session was expired. Please signin again!", 'Msis-Web');
+                AuthenService.clearCredentials();
+                $location.path('/signin');
+            } else {
+                return $http.delete(uri + '/' + id).then(successCallback, errorCallback);
+            }
+            return false;
         }
 
         function successCallback(response) {
